Adds The Chilldown, Cargo Luxury
A ship designed for the cold vacuum of space is thematically designed after that same vacuum, running with cold freezer floors and a pretty wicked air conditioning system. Of course, you'll have to turn it on - Complete with beds devoid of sheets ontop of cold tile floors, this chilling experience is sure to freeze any cargo technician out. Works ideally as a planetary cargo lander with a cargo bay, buildable conveyor, and express console designed to buy (but not sell!) from NT Centcomm and build up the rest of your bases' needs.